Optimizations

+++ Cogs/Fun.py
@@ -1,9 +1,6 @@
 from discord.ext import commands
-from Cogs.Tools import JsonTools, MessageTools
+from Cogs.Tools import MessageTools
 import discord
-import asyncio
-import random
-from itertools import cycle
 
 
 class Fun(commands.Cog):
@@ -18,22 +15,31 @@ class Fun(commands.Cog):
             for m in self.bot.get_guild(477829362771689484).emojis:
                 if m.name == 'BoiGif':
                     if arg == self.bot.user:
+                        await ctx.message.add_reaction('🇧')
+                        await ctx.message.add_reaction('🇴')
+                        await ctx.message.add_reaction('🇮')
                         await ctx.message.add_reaction(m)
-                        await MessageTools.sendSimpleEmbed(ctx, f'{ctx.author.mention}, You roast me I roast you',
-                                                           delete=False)
+                        await MessageTools.sendSimpleEmbed(ctx, f'{ctx.author.mention}, You roast me I roast you', delete=False)
                     elif arg == ctx.author:
+                        await ctx.message.add_reaction('🇧')
+                        await ctx.message.add_reaction('🇴')
+                        await ctx.message.add_reaction('🇮')
                         await ctx.message.add_reaction(m)
                     else:
                         async for message in ctx.channel.history(limit=50):
                             if message.author == arg:
+                                await message.add_reaction('🇧')
+                                await message.add_reaction('🇴')
+                                await message.add_reaction('🇮')
                                 await message.add_reaction(m)
                                 break
 
     @commands.command()
     async def purge(self, ctx, num: int):
         if MessageTools.correct_command_use(ctx, mod_command=True):
-            await ctx.message.delete()
-            await ctx.channel.purge(limit=num)
+            if ctx.message.author != self.bot.user:
+                await ctx.message.delete()
+                await ctx.channel.purge(limit=num)
 
 
 def setup(bot):
